[CP-2474] Move Loader Display Logic to Custom Hook
